Macro \I works by line; supports count / range (ref 3599bf9)
Previous version works on a whole paragraph.
This revision allows it to work on a single line at a time,
with support for optional count or range, for multiple lines.

This is also to make \I work in the same way as \O

fyi, function for \I renamed: Extract_first_args_para() -> Extract_first_arg()