commentor: restore cursor to 'opfunc's, for normal mode maps
Builds on prev commit.
Extends restoring cursor feature
from visual mode to normal mode maps.

The quirk is that storing of cursor / window happens in the mapping,
_before_ the operator;
but the restoring of cursor / window happens at the end, _within_ the operator.

As an aside, the normal mode commenting maps are dot-repeatable :)