include parameters (oma, mar) in r's mfrow snippet
with multiple plots `par(mfrow=c(.,.))`,
the margins for each individual subplot (default: c(5,4,1,1))
results in a lot of white space between subplots, and
each subplot is too small.

The remedy is to modify the outer plot margins (oma),
and the subplot margins (mar).
These are modified as follows:
  oma = c(5,4,0,0) + 0.1  (original default: c(0,0,0,0)), and
  mar = c(0,0,1,1) + 0.1  (original default: c(5,4,1,1)).

cf. 3e1bddf